Fix OoO deadlock on real ChampSim traces
Two bugs in the OoO core's register handling, both hidden by synth
traces. Real SPEC2017 traces deadlocked or crashed on full mode; cache
mode worked. Found via AddressSanitizer.

1. RAT was sized for project2 (kNumArchRegs = 32) and used int8_t for
   register IDs. ChampSim records emit uint8_t (0-255) — IDs above 31
   read past the RAT, IDs above 127 sign-wrap to a wild address.
   Fix: kNumArchRegs 32 -> 256; widen Inst::dest/src1/src2,
   RobEntry::dest_reg, and Rat::read/write_use/mark_complete from
   int8_t to int16_t.

2. LSU store completion set ROB.ready and erased the schedQ entry but
   never called rat_.mark_complete or sq_.wake_dependents. Fine for
   project2/synth (stores have inst.dest == kNoReg), wrong for real
   traces where stores routinely have a register destination
   (push/pop, ARM pre-/post-indexed sp writeback, lock cmpxchg flag
   updates). Younger ops reading that register sat at src_ready=false
   forever and the pipeline deadlocked.
   Fix: store path calls writeback() unconditionally; safe for
   kNoReg dest because mark_complete/wake_dependents no-op.

Also: strip inline comments from --trace-list manifests in
full_mode.cpp so the existing path-with-trailing-comment convention
in traces/mixes/*.txt actually works.

Verification: 128/128 ctest still passes (no synth regression). Real
traces now produce published-baseline numbers under baseline.json:
  mcf       IPC 0.446  L1 miss 10.30%
  perlbench IPC 1.287  L1 miss  0.04%
  leela     IPC 1.210  L1 miss  0.56%
  xz        IPC 1.538  L1 miss  2.12%

Full investigation log + walkthroughs in
report_doc/16-real-trace-deadlock-rootcause.md.

+++ include/comparch/ooo/inst.hpp
@@ -35,17 +35,22 @@ enum class Opcode : std::uint8_t {
     Branch = 4,
 };
 
-// Sentinel for "no operand". Matches project2's int8_t -1 convention.
-inline constexpr std::int8_t kNoReg = -1;
+// Sentinel for "no operand". Widened to int16_t so we can hold the full
+// uint8_t range of ChampSim register IDs (0-255) plus -1 for kNoReg.
+// Project2 used int8_t with NUM_REGS=32; real ChampSim records emit
+// register IDs above 127 (DynamoRIO dr_reg_id_t) which would wrap to
+// negative under int8_t and trigger out-of-bounds reads in the RAT.
+inline constexpr std::int16_t kNoReg = -1;
 
 struct Inst {
     std::uint64_t pc            = 0;
     Opcode        opcode        = Opcode::Alu;
 
-    // Architectural register numbers. -1 == no operand.
-    std::int8_t   dest          = kNoReg;
-    std::int8_t   src1          = kNoReg;
-    std::int8_t   src2          = kNoReg;
+    // Architectural register numbers. -1 == no operand. Widened to
+    // int16_t to cover ChampSim's full 0-255 register ID range.
+    std::int16_t  dest          = kNoReg;
+    std::int16_t  src1          = kNoReg;
+    std::int16_t  src2          = kNoReg;
 
     // Memory address for loads / stores. 0 if not a memory op.
     std::uint64_t mem_addr      = 0;
